Move AccountSyncedSettingsInternal, AccountSyncedSettings, and AccountSettings to commons
- Create expect/actual PlatformLocale for getDeviceLanguages() and getDefaultLanguage()
  - Android: uses Resources.getSystem().configuration.locales (API 26+)
  - JVM: uses java.util.Locale.getDefault()
  - iOS: uses NSLocale.preferredLanguages
- Move AccountSyncedSettingsInternal and all its data classes to commons/commonMain
- Move AccountSyncedSettings and preference classes to commons/commonMain
- Move AccountSettings, TopFilter, and related types to commons/commonMain
- Replace java.util.UUID with kotlin.uuid.Uuid for KMP compatibility
- Add @Contextual annotation for Address serialization in TopFilter
- Leave typealiases in amethyst module for backward compatibility
- Update TopFilter imports across amethyst module to point to commons
