+++ amethyst/src/main/java/com/vitorpamplona/amethyst/LocalPreferences.kt
@@ -25,8 +25,8 @@ import android.content.Context
 import android.content.SharedPreferences
 import androidx.compose.runtime.Immutable
 import androidx.core.content.edit
+import com.vitorpamplona.amethyst.commons.model.TopFilter
 import com.vitorpamplona.amethyst.model.AccountSettings
-import com.vitorpamplona.amethyst.model.TopFilter
 import com.vitorpamplona.amethyst.model.UiSettings
 import com.vitorpamplona.amethyst.model.nip47WalletConnect.NwcWalletEntry
 import com.vitorpamplona.amethyst.model.nip47WalletConnect.NwcWalletEntryNorm
