+++ amethyst/src/main/java/com/vitorpamplona/amethyst/model/Account.kt
@@ -2445,8 +2445,9 @@ class Account(
         scope.launch(Dispatchers.IO) {
             @OptIn(kotlinx.coroutines.FlowPreview::class)
             settings.saveable.debounce(1000).collect {
-                if (it.accountSettings != null) {
-                    LocalPreferences.saveToEncryptedStorage(it.accountSettings)
+                val accountSettings = it.accountSettings
+                if (accountSettings != null) {
+                    LocalPreferences.saveToEncryptedStorage(accountSettings)
                 }
             }
         }
