+++ commons/src/commonMain/kotlin/com/vitorpamplona/amethyst/commons/platform/PlatformLocale.kt
@@ -18,11 +18,8 @@
  * AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM, OUT OF OR IN CONNECTION
  * WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE SOFTWARE.
  */
-package com.vitorpamplona.amethyst.model
+package com.vitorpamplona.amethyst.commons.platform
 
-// Re-export all types from commons for backward compatibility
-typealias AccountSyncedSettings = com.vitorpamplona.amethyst.commons.model.AccountSyncedSettings
-typealias AccountReactionPreferences = com.vitorpamplona.amethyst.commons.model.AccountReactionPreferences
-typealias AccountZapPreferences = com.vitorpamplona.amethyst.commons.model.AccountZapPreferences
-typealias AccountLanguagePreferences = com.vitorpamplona.amethyst.commons.model.AccountLanguagePreferences
-typealias AccountSecurityPreferences = com.vitorpamplona.amethyst.commons.model.AccountSecurityPreferences
+expect fun getDeviceLanguages(): Set<String>
+
+expect fun getDefaultLanguage(): String
